requirements.txt
altair==6.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
decorator==5.2.1
dotenv==0.9.9
exceptiongroup==1.3.1
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
ipython==8.39.0
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
minio==7.2.20
narwhals==2.17.0
networkx==3.4.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
prompt_toolkit==3.0.52
protobuf==6.33.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pycryptodome==3.23.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
pyvis==0.3.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.54.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0